refactor: extract workspace metadata reconciliation
