+++ inc/Workspace/Workspace.php
@@ -21,10 +21,12 @@ use DataMachineCode\Storage\WorktreeInventoryRepository;
 defined( 'ABSPATH' ) || exit;
 
 require_once __DIR__ . '/WorkspaceArtifactCleanup.php';
+require_once __DIR__ . '/WorkspaceMetadataReconciliation.php';
 
 class Workspace {
 
 	use WorkspaceArtifactCleanup;
+	use WorkspaceMetadataReconciliation;
 
 	/**
 	 * Maximum file size for reading (1 MB).
@@ -3801,1204 +3803,6 @@ class Workspace {
 				'metadata'    => $metadata,
 			),
 			$disk_fields
-		);
-	}
-
-	/**
-	 * Reconcile lifecycle metadata for unmanaged worktrees without removing anything.
-	 *
-	 * Dry-runs build a reviewed plan from the current git worktree listing.
-	 * Passing `limit` and/or `offset` bounds expensive per-worktree probes to
-	 * only that page; omitting both preserves the historical full scan.
-	 * Applying a plan revalidates handle/path/repo/branch before writing metadata.
-	 *
-	 * @param array $opts Options: dry_run bool, apply_plan array, limit int, offset int, until_budget string.
-	 * @return array<string,mixed>|\WP_Error
-	 */
-	public function worktree_reconcile_metadata( array $opts = array() ): array|\WP_Error {
-		$started_at = microtime( true );
-		$dry_run      = ! empty( $opts['dry_run'] );
-		$apply        = ! empty( $opts['apply'] );
-		$via_jobs     = ! empty( $opts['via_jobs'] );
-		$source       = isset( $opts['source'] ) ? trim( (string) $opts['source'] ) : 'workspace_metadata_reconcile';
-		$apply_plan   = isset( $opts['apply_plan'] ) && is_array( $opts['apply_plan'] ) ? $opts['apply_plan'] : null;
-		$until_budget = isset( $opts['until_budget'] ) ? trim( (string) $opts['until_budget'] ) : '';
-		$paged        = array_key_exists( 'limit', $opts ) || array_key_exists( 'offset', $opts ) || '' !== $until_budget;
-		$limit        = $paged ? ( array_key_exists( 'limit', $opts ) ? (int) $opts['limit'] : self::METADATA_RECONCILE_DEFAULT_LIMIT ) : 0;
-		$offset       = $paged ? max( 0, (int) ( $opts['offset'] ?? 0 ) ) : 0;
-
-		if ( null !== $apply_plan ) {
-			return $this->apply_worktree_metadata_reconciliation_plan( $apply_plan );
-		}
-
-		if ( ! $dry_run && ! $apply ) {
-			return new \WP_Error( 'metadata_reconcile_requires_review', 'Metadata reconciliation is dry-run-first. Pass --dry-run to review JSON output, or pass apply=true for DMC-owned lifecycle reconciliation.', array( 'status' => 400 ) );
-		}
-		if ( $via_jobs && ( $dry_run || ! $apply || null !== $apply_plan ) ) {
-			return new \WP_Error( 'metadata_reconcile_via_jobs_requires_apply', 'Job-backed metadata reconciliation requires apply=true without dry_run or apply_plan.', array( 'status' => 400 ) );
-		}
-		if ( $via_jobs && '' !== $until_budget ) {
-			return new \WP_Error( 'metadata_reconcile_budget_via_jobs_unsupported', 'Metadata reconciliation --until-budget cannot be combined with --via-jobs.', array( 'status' => 400 ) );
-		}
-		if ( $paged && $limit <= 0 ) {
-			return new \WP_Error( 'invalid_metadata_reconcile_limit', 'Metadata reconciliation --limit must be greater than 0.', array( 'status' => 400 ) );
-		}
-		if ( '' !== $until_budget ) {
-			$budget_seconds = $this->parse_worktree_metadata_reconciliation_budget( $until_budget );
-			if ( is_wp_error( $budget_seconds ) ) {
-				return $budget_seconds;
-			}
-
-			if ( $apply && ! $dry_run ) {
-				return $this->drain_worktree_metadata_reconciliation_budget( $limit, $offset, $until_budget, $budget_seconds );
-			}
-		}
-		if ( $via_jobs && ! $paged ) {
-			$limit  = self::METADATA_RECONCILE_DEFAULT_LIMIT;
-			$offset = 0;
-			$paged  = true;
-		}
-
-		$listing = $this->worktree_list(
-			null,
-			null,
-			$paged ? array(
-				'include_status' => false,
-				'include_disk'   => false,
-			) : array()
-		);
-		if ( is_wp_error( $listing ) ) {
-			return $listing;
-		}
-
-		$all_worktrees = array_values( array_filter(
-			(array) ( $listing['worktrees'] ?? array() ),
-			fn( $wt ) => empty( $wt['is_primary'] )
-		) );
-		$total_worktrees = count( $all_worktrees );
-		$page_worktrees  = $paged ? array_slice( $all_worktrees, $offset, $limit ) : $all_worktrees;
-
-		$proposals    = array();
-		$skipped      = array();
-		$github_cache = array();
-		$fetched      = array();
-
-		$budget_context = $this->build_worktree_loop_budget_context( $opts, $started_at );
-		$budget_stopped = false;
-		foreach ( $page_worktrees as $index => $wt ) {
-			if ( null !== $budget_context && $this->is_worktree_loop_budget_exhausted( $budget_context ) ) {
-				$budget_stopped = true;
-				$page_worktrees = array_slice( $page_worktrees, 0, $index );
-				break;
-			}
-			$row_started = microtime( true );
-			$proposal = $this->build_worktree_metadata_reconciliation_row( $wt, $github_cache, $fetched );
-			$elapsed_ms = (int) round( ( microtime( true ) - $row_started ) * 1000 );
-			if ( isset( $proposal['proposal'] ) ) {
-				$proposal['proposal']['elapsed_ms'] = $elapsed_ms;
-				$proposals[] = $proposal['proposal'];
-			} elseif ( isset( $proposal['skip'] ) ) {
-				$proposal['skip']['elapsed_ms'] = $elapsed_ms;
-				$skipped[] = $proposal['skip'];
-			}
-		}
-
-		$classified_skips = $this->classify_worktree_metadata_reconciliation_skips( $skipped );
-		$pagination       = $paged ? $this->build_worktree_metadata_reconciliation_pagination( $total_worktrees, count( $page_worktrees ), $limit, $offset ) : null;
-		if ( null !== $pagination && $budget_stopped ) {
-			$pagination['partial']      = true;
-			$pagination['complete']     = false;
-			$pagination['next_offset']  = $offset + count( $page_worktrees );
-			$pagination['next_command'] = sprintf( 'studio wp datamachine-code workspace worktree reconcile-metadata --%s --limit=%d --offset=%d%s --format=json', $apply ? 'apply' : 'dry-run', $limit, (int) $pagination['next_offset'], null !== $budget_context ? ' --until-budget=' . (string) $budget_context['label'] : '' );
-		}
-
-		$plan = array(
-			'success'        => true,
-			'dry_run'        => $dry_run,
-			'applied'        => false,
-			'generated_at'   => gmdate( 'c' ),
-			'workspace_path' => $this->workspace_path,
-			'proposals'      => $proposals,
-			'written'        => array(),
-			'skipped'        => $skipped,
-			'still_unsafe'      => $classified_skips['still_unsafe'],
-			'external_worktrees' => $classified_skips['external_worktrees'],
-			'summary'        => $this->build_worktree_metadata_reconciliation_summary( $paged ? count( $page_worktrees ) : count( (array) ( $listing['worktrees'] ?? array() ) ), $proposals, array(), $skipped ),
-		);
-		if ( null !== $pagination ) {
-			$plan['pagination'] = $pagination;
-			$plan['evidence']   = array(
-				'scope'  => 'paginated metadata reconciliation dry-run',
-				'note'   => 'Only this page ran per-worktree dirty, unpushed, merge-signal, and GitHub probes. Run the next_offset page until complete for full inventory review.',
-				'fields_skipped_by_listing' => (array) ( $listing['fields_skipped'] ?? array() ),
-			);
-			if ( null !== $budget_context ) {
-				$plan['evidence']['budget'] = $this->summarize_worktree_loop_budget_context( $budget_context, $budget_stopped );
-			}
-		}
-
-		if ( $apply ) {
-			if ( $via_jobs ) {
-				return $this->schedule_worktree_metadata_reconciliation_pages( $plan, $limit, $source, $started_at );
-			}
-			$plan['direct_apply'] = true;
-			return $this->apply_worktree_metadata_reconciliation_plan( $plan );
-		}
-
-		return $plan;
-	}
-
-	/**
-	 * Drain paged direct metadata reconciliation until the time budget is nearly exhausted.
-	 *
-	 * @param int    $limit          Page size.
-	 * @param int    $offset         Starting offset.
-	 * @param string $budget_label   Original compact budget label.
-	 * @param int    $budget_seconds Parsed budget in seconds.
-	 * @return array<string,mixed>|\WP_Error
-	 */
-	private function drain_worktree_metadata_reconciliation_budget( int $limit, int $offset, string $budget_label, int $budget_seconds ): array|\WP_Error {
-		$started_at      = microtime( true );
-		$reserve_seconds = min( 5.0, max( 1.0, $budget_seconds * 0.1 ) );
-		$pages           = array();
-		$proposals       = array();
-		$written         = array();
-		$skipped         = array();
-		$scanned         = 0;
-		$next_offset     = $offset;
-		$last_pagination = null;
-
-		do {
-			$page = $this->worktree_reconcile_metadata(
-				array(
-					'apply'                   => true,
-					'limit'                   => $limit,
-					'offset'                  => $next_offset,
-					'internal_budget_label'   => $budget_label,
-					'internal_budget_seconds' => $budget_seconds,
-					'internal_budget_started' => $started_at,
-				)
-			);
-			if ( is_wp_error( $page ) ) {
-				return $page;
-			}
-
-			$page_pagination = (array) ( $page['pagination'] ?? array() );
-			$page_scanned    = (int) ( $page_pagination['scanned'] ?? 0 );
-			$last_pagination = $page_pagination;
-			$scanned        += $page_scanned;
-			$proposals       = array_merge( $proposals, (array) ( $page['proposals'] ?? array() ) );
-			$written         = array_merge( $written, (array) ( $page['written'] ?? array() ) );
-			$skipped         = array_merge( $skipped, (array) ( $page['skipped'] ?? array() ) );
-			$pages[]         = array(
-				'offset'   => (int) ( $page_pagination['offset'] ?? $next_offset ),
-				'limit'    => (int) ( $page_pagination['limit'] ?? $limit ),
-				'scanned'  => $page_scanned,
-				'written'  => count( (array) ( $page['written'] ?? array() ) ),
-				'skipped'  => count( (array) ( $page['skipped'] ?? array() ) ),
-				'complete' => (bool) ( $page_pagination['complete'] ?? false ),
-			);
-
-			if ( ! empty( $page_pagination['complete'] ) || null === ( $page_pagination['next_offset'] ?? null ) || $page_scanned <= 0 ) {
-				break;
-			}
-
-			$next_offset = (int) $page_pagination['next_offset'];
-		} while ( ( $budget_seconds - ( microtime( true ) - $started_at ) ) > $reserve_seconds );
-
-		$elapsed      = microtime( true ) - $started_at;
-		$complete     = ! empty( $last_pagination['complete'] );
-		$partial      = ! $complete;
-		$next_command = $partial ? sprintf(
-			'studio wp datamachine-code workspace worktree reconcile-metadata --apply --limit=%d --offset=%d --until-budget=%s --format=json',
-			$limit,
-			(int) ( $last_pagination['next_offset'] ?? $next_offset ),
-			$budget_label
-		) : null;
-
-		$classified_skips = $this->classify_worktree_metadata_reconciliation_skips( $skipped );
-		$pagination       = array(
-			'total'       => (int) ( $last_pagination['total'] ?? 0 ),
-			'offset'      => $offset,
-			'limit'       => $limit,
-			'scanned'     => $scanned,
-			'partial'     => $partial,
-			'complete'    => $complete,
-			'next_offset' => $partial ? (int) ( $last_pagination['next_offset'] ?? $next_offset ) : null,
-		);
-		if ( null !== $next_command ) {
-			$pagination['next_command'] = $next_command;
-		}
-
-		return array(
-			'success'        => true,
-			'dry_run'        => false,
-			'applied'        => true,
-			'direct_apply'   => true,
-			'generated_at'   => gmdate( 'c' ),
-			'workspace_path' => $this->workspace_path,
-			'proposals'      => $proposals,
-			'written'        => $written,
-			'skipped'        => $skipped,
-			'still_unsafe'      => $classified_skips['still_unsafe'],
-			'external_worktrees' => $classified_skips['external_worktrees'],
-			'summary'        => $this->build_worktree_metadata_reconciliation_summary( $scanned, $proposals, $written, $skipped ),
-			'pagination'     => $pagination,
-			'evidence'       => array_filter(
-				array(
-					'scope'                   => 'time-budgeted metadata reconciliation direct apply',
-					'apply_source'            => 'direct_apply',
-					'budget'                  => $budget_label,
-					'budget_seconds'          => $budget_seconds,
-					'reserve_seconds'         => $reserve_seconds,
-					'elapsed_seconds'         => round( $elapsed, 3 ),
-					'budget_nearly_exhausted' => $partial,
-					'pages'                   => $pages,
-					'next_command'            => $next_command,
-				)
-			),
-		);
-	}
-
-	/**
-	 * Schedule bounded metadata reconciliation apply pages as system jobs.
-	 *
-	 * @param array<string,mixed> $first_page First bounded dry-run page.
-	 * @param int                 $limit      Page size.
-	 * @param string              $source     Caller marker.
-	 * @param float               $started_at Start timestamp.
-	 * @return array<string,mixed>|\WP_Error
-	 */
-	private function schedule_worktree_metadata_reconciliation_pages( array $first_page, int $limit, string $source, float $started_at ): array|\WP_Error {
-		if ( ! class_exists( '\DataMachine\Engine\Tasks\TaskScheduler' ) ) {
-			return new \WP_Error( 'task_scheduler_unavailable', 'Data Machine TaskScheduler is unavailable; cannot schedule metadata reconciliation jobs.', array( 'status' => 500 ) );
-		}
-
-		$pagination = (array) ( $first_page['pagination'] ?? array() );
-		$total      = max( 0, (int) ( $pagination['total'] ?? $pagination['scanned'] ?? 0 ) );
-		$start      = max( 0, (int) ( $pagination['offset'] ?? 0 ) );
-		$limit      = max( 1, $limit );
-		$items      = array();
-		for ( $offset = $start; $offset < $total; $offset += $limit ) {
-			$items[] = array(
-				'chunk_type'  => 'metadata_reconciliation_page',
-				'chunk_index' => count( $items ),
-				'limit'       => $limit,
-				'offset'      => $offset,
-				'source'      => $source,
-			);
-		}
-
-		if ( array() === $items ) {
-			return array(
-				'success'        => true,
-				'dry_run'        => false,
-				'applied'        => false,
-				'job_backed'     => true,
-				'generated_at'   => gmdate( 'c' ),
-				'workspace_path' => $this->workspace_path,
-				'proposals'      => array(),
-				'written'        => array(),
-				'skipped'        => array(),
-				'still_unsafe'       => array(),
-				'external_worktrees' => array(),
-				'summary'            => array(
-					'inspected'      => 0,
-					'proposed'       => 0,
-					'written'        => 0,
-					'skipped'        => 0,
-					'scheduled_jobs' => 0,
-					'limit'          => $limit,
-				),
-				'pagination'     => $pagination,
-				'evidence'       => array(
-					'elapsed_ms' => (int) round( ( microtime( true ) - $started_at ) * 1000 ),
-					'note'       => 'No metadata reconciliation pages eligible for scheduling.',
-					'source'     => $source,
-				),
-			);
-		}
-
-		$batch_result = \DataMachine\Engine\Tasks\TaskScheduler::scheduleBatch(
-			'worktree_cleanup_chunk',
-			$items,
-			array( 'source' => $source )
-		);
-
-		if ( false === $batch_result ) {
-			return new \WP_Error( 'metadata_reconcile_schedule_failed', 'Failed to schedule metadata reconciliation page jobs.', array( 'status' => 500 ) );
-		}
-
-		return array(
-			'success'        => true,
-			'dry_run'        => false,
-			'applied'        => false,
-			'job_backed'     => true,
-			'generated_at'   => gmdate( 'c' ),
-			'workspace_path' => $this->workspace_path,
-			'proposals'      => array_values( (array) ( $first_page['proposals'] ?? array() ) ),
-			'written'        => array(),
-			'skipped'        => array(),
-			'still_unsafe'       => array_values( (array) ( $first_page['still_unsafe'] ?? array() ) ),
-			'external_worktrees' => array_values( (array) ( $first_page['external_worktrees'] ?? array() ) ),
-			'summary'            => array_merge(
-				(array) ( $first_page['summary'] ?? array() ),
-				array(
-					'written'        => 0,
-					'scheduled_jobs' => count( $items ),
-					'limit'          => $limit,
-				)
-			),
-			'pagination'     => $pagination,
-			'evidence'       => array(
-				'elapsed_ms'     => (int) round( ( microtime( true ) - $started_at ) * 1000 ),
-				'scope'          => 'job-backed metadata reconciliation apply',
-				'page_offsets'   => array_column( $items, 'offset' ),
-				'batch_job_id'   => (int) ( $batch_result['batch_job_id'] ?? 0 ),
-				'direct_job_ids' => $batch_result['job_ids'] ?? array(),
-				'source'         => $source,
-			),
-		);
-	}
-
-	/**
-	 * Parse a compact metadata reconciliation time budget.
-	 *
-	 * @param string $duration Duration like 60s, 10m, or 1h.
-	 * @return int|\WP_Error Seconds on success.
-	 */
-	private function parse_worktree_metadata_reconciliation_budget( string $duration ): int|\WP_Error {
-		if ( ! preg_match( '/^(\d+)([smh])$/', trim( $duration ), $matches ) ) {
-			return new \WP_Error( 'invalid_metadata_reconcile_budget', 'Invalid --until-budget duration. Use a compact value like 60s, 10m, or 1h.', array( 'status' => 400 ) );
-		}
-
-		$value = (int) $matches[1];
-		if ( $value <= 0 ) {
-			return new \WP_Error( 'invalid_metadata_reconcile_budget', 'Invalid --until-budget duration. Duration must be greater than zero.', array( 'status' => 400 ) );
-		}
-
-		$unit_seconds = array(
-			's' => 1,
-			'm' => 60,
-			'h' => 3600,
-		);
-
-		return $value * $unit_seconds[ $matches[2] ];
-	}
-
-	/**
-	 * Build a shared wall-clock budget context for expensive worktree loops.
-	 *
-	 * @param array<string,mixed> $opts       Operation options.
-	 * @param float               $started_at Operation start timestamp.
-	 * @return array<string,mixed>|null
-	 */
-	private function build_worktree_loop_budget_context( array $opts, float $started_at ): ?array {
-		$label = isset( $opts['internal_budget_label'] ) ? trim( (string) $opts['internal_budget_label'] ) : ( isset( $opts['until_budget'] ) ? trim( (string) $opts['until_budget'] ) : '' );
-		if ( '' === $label && ! isset( $opts['internal_budget_seconds'] ) ) {
-			return null;
-		}
-
-		$seconds = isset( $opts['internal_budget_seconds'] ) ? (int) $opts['internal_budget_seconds'] : $this->parse_worktree_metadata_reconciliation_budget( $label );
-		if ( is_wp_error( $seconds ) || $seconds <= 0 ) {
-			return null;
-		}
-
-		$started = isset( $opts['internal_budget_started'] ) ? (float) $opts['internal_budget_started'] : $started_at;
-		$reserve = min( 5.0, max( 0.1, $seconds * 0.1 ) );
-
-		return array(
-			'label'           => '' === $label ? $seconds . 's' : $label,
-			'seconds'         => $seconds,
-			'started_at'      => $started,
-			'reserve_seconds' => $reserve,
-		);
-	}
-
-	/**
-	 * Determine whether an expensive loop should stop before another row starts.
-	 *
-	 * @param array<string,mixed> $context Budget context.
-	 * @return bool
-	 */
-	private function is_worktree_loop_budget_exhausted( array $context ): bool {
-		$remaining = (float) ( $context['seconds'] ?? 0 ) - ( microtime( true ) - (float) ( $context['started_at'] ?? microtime( true ) ) );
-		return $remaining <= (float) ( $context['reserve_seconds'] ?? 1.0 );
-	}
-
-	/**
-	 * Summarize budget evidence for JSON responses.
-	 *
-	 * @param array<string,mixed> $context   Budget context.
-	 * @param bool                $exhausted Whether the loop stopped for budget.
-	 * @return array<string,mixed>
-	 */
-	private function summarize_worktree_loop_budget_context( array $context, bool $exhausted ): array {
-		$elapsed = max( 0.0, microtime( true ) - (float) ( $context['started_at'] ?? microtime( true ) ) );
-		return array(
-			'label'             => (string) ( $context['label'] ?? '' ),
-			'budget_seconds'    => (int) ( $context['seconds'] ?? 0 ),
-			'reserve_seconds'   => (float) ( $context['reserve_seconds'] ?? 0 ),
-			'elapsed_seconds'   => round( $elapsed, 3 ),
-			'budget_exhausted'  => $exhausted,
-			'remaining_seconds' => round( max( 0.0, (float) ( $context['seconds'] ?? 0 ) - $elapsed ), 3 ),
-		);
-	}
-
-	/**
-	 * Build one metadata reconciliation row for a current worktree listing row.
-	 *
-	 * @param array<string,mixed> $wt Worktree list row.
-	 * @return array{proposal?:array<string,mixed>,skip?:array<string,mixed>}
-	 */
-	private function build_worktree_metadata_reconciliation_row( array $wt, array &$github_cache = array(), array &$fetched = array() ): array {
-		$handle   = (string) ( $wt['handle'] ?? '' );
-		$repo     = (string) ( $wt['repo'] ?? '' );
-		$branch   = (string) ( $wt['branch'] ?? '' );
-		$path     = (string) ( $wt['path'] ?? '' );
-		$metadata = is_array( $wt['metadata'] ?? null ) ? (array) $wt['metadata'] : array();
-
-		$base_row = array(
-			'handle'   => $handle,
-			'repo'     => $repo,
-			'branch'   => $branch,
-			'path'     => $path,
-			'metadata' => array() === $metadata ? null : $metadata,
-		);
-
-		if ( ! empty( $wt['external'] ) ) {
-			return array(
-				'skip' => array_merge(
-					$base_row,
-					array(
-						'reason_code' => 'external_worktree',
-						'reason'      => 'external worktree outside the DMC workspace',
-					)
-				),
-			);
-		}
-
-		$missing = array_values( array_filter( array(
-			'' === $handle ? 'handle' : '',
-			'' === $repo ? 'repo' : '',
-			'' === $branch ? 'branch' : '',
-			'' === $path ? 'path' : '',
-		) ) );
-		if ( array() !== $missing ) {
-			return array(
-				'skip' => array_merge(
-					$base_row,
-					array(
-						'reason_code'    => 'missing_identity',
-						'reason'         => 'current worktree row is missing required identity fields',
-						'missing_fields' => $missing,
-					)
-				),
-			);
-		}
-
-		$parsed = $this->parse_handle( $handle );
-		if ( empty( $parsed['is_worktree'] ) || $parsed['repo'] !== $repo ) {
-			return array(
-				'skip' => array_merge(
-					$base_row,
-					array(
-						'reason_code' => 'noncanonical_handle',
-						'reason'      => 'worktree is not represented by a canonical <repo>@<slug> workspace handle',
-					)
-				),
-			);
-		}
-
-		$dirty = $wt['dirty'] ?? null;
-		if ( null === $dirty ) {
-			$dirty = $this->probe_worktree_dirty_count( $path, self::CLEANUP_GIT_PROBE_TIMEOUT );
-			if ( is_wp_error( $dirty ) ) {
-				return array(
-					'skip' => array_merge(
-						$base_row,
-						array(
-							'reason_code' => 'probe_timeout',
-							'reason'      => 'dirty-state probe failed - leaving lifecycle unchanged: ' . $dirty->get_error_message(),
-						)
-					),
-				);
-			}
-		}
-		$dirty   = (int) $dirty;
-		$unpushed = $this->count_unpushed_commits( $path );
-		if ( is_wp_error( $unpushed ) ) {
-			return array(
-				'skip' => array_merge(
-					$base_row,
-					array(
-						'reason_code' => 'probe_timeout',
-						'reason'      => 'cleanup safety probe failed - leaving lifecycle unchanged: ' . $unpushed->get_error_message(),
-					)
-				),
-			);
-		}
-
-		$finalizer_signal = $this->detect_worktree_lifecycle_finalizer_signal( $wt, $metadata, $github_cache, $fetched );
-		if ( null !== $finalizer_signal && 'probe-timeout' === ( $finalizer_signal['signal'] ?? '' ) ) {
-			return array(
-				'skip' => array_merge(
-					$base_row,
-					array(
-						'reason_code' => 'probe_timeout',
-						'reason'      => 'merge-signal probe timed out - leaving lifecycle unchanged: ' . $finalizer_signal['reason'],
-					)
-				),
-			);
-		}
-
-		if ( null !== $finalizer_signal && ! $this->has_explicit_cleanup_eligible_state( $metadata ) ) {
-			if ( $dirty > 0 || $unpushed > 0 ) {
-				return array(
-					'skip' => array_merge(
-						$base_row,
-						array(
-							'reason_code' => 'unsafe_cleanup_eligible_state',
-							'reason'      => 'merged lifecycle signal found, but dirty or unpushed worktree is not auto-finalized',
-							'dirty'       => $dirty,
-							'unpushed'    => $unpushed,
-							'signal'      => $finalizer_signal['signal'],
-						)
-					),
-				);
-			}
-
-			$finalized_state    = (string) ( $finalizer_signal['finalized_state'] ?? WorktreeContextInjector::STATE_MERGED );
-			$finalizer_metadata = WorktreeContextInjector::build_finalizer_metadata(
-				$finalized_state,
-				isset( $finalizer_signal['pr_url'] ) ? (string) $finalizer_signal['pr_url'] : null
-			);
-			$evidence           = array_filter(
-				array(
-					'signal'          => $finalizer_signal['signal'],
-					'finalized_state' => $finalized_state,
-					'reason'          => $finalizer_signal['reason'],
-					'detected_at'     => gmdate( 'c' ),
-					'dirty'           => $dirty,
-					'unpushed'        => $unpushed,
-					'pr_url'          => $finalizer_signal['pr_url'] ?? null,
-				),
-				fn( $value ) => null !== $value && '' !== $value
-			);
-			$proposed           = array_merge(
-				$metadata,
-				array(
-					'handle'       => $handle,
-					'repo'         => $repo,
-					'branch'       => $branch,
-					'path'         => $path,
-					'observed_at'  => gmdate( 'c' ),
-					'last_seen_at' => gmdate( 'c' ),
-				),
-				$finalizer_metadata,
-				array(
-					'auto_finalized_by'            => 'worktree_reconcile_metadata',
-					'auto_finalized_signal'        => $finalizer_signal['signal'],
-					'auto_finalized_reason'        => $finalizer_signal['reason'],
-					'cleanup_eligibility_evidence' => $evidence,
-				)
-			);
-
-			if ( empty( $proposed['created_at'] ) ) {
-				$created_at = file_exists( $path ) ? filemtime( $path ) : false;
-				if ( false !== $created_at ) {
-					$proposed['created_at'] = gmdate( 'c', (int) $created_at );
-				}
-			}
-
-			return array(
-				'proposal' => array_merge(
-					$base_row,
-					array(
-						'reason_code'       => 'auto_finalize_merged',
-						'reason'            => 'merged PR or branch state proves lifecycle can be finalized as cleanup_eligible',
-						'dirty'             => $dirty,
-						'unpushed'          => $unpushed,
-						'signal'            => $finalizer_signal['signal'],
-						'pr_url'            => $finalizer_signal['pr_url'] ?? null,
-						'proposed_metadata' => $proposed,
-						'source_map'        => array(
-							'handle'                       => 'filesystem',
-							'repo'                         => 'filesystem',
-							'branch'                       => 'git',
-							'path'                         => 'git',
-							'created_at'                   => empty( $metadata['created_at'] ) ? 'filesystem' : 'metadata',
-							'observed_at'                  => 'reconcile_run',
-							'lifecycle_state'              => 'merge_signal',
-							'finalized_state'              => 'merge_signal',
-							'cleanup_eligibility_evidence' => 'merge_signal',
-						),
-					),
-				),
-			);
-		}
-
-		$proposed   = $metadata;
-		$source_map = array();
-		$this->set_reconciled_metadata_field( $proposed, $source_map, 'handle', $handle, 'filesystem' );
-		$this->set_reconciled_metadata_field( $proposed, $source_map, 'repo', $repo, 'filesystem' );
-		$this->set_reconciled_metadata_field( $proposed, $source_map, 'branch', $branch, 'git' );
-		$this->set_reconciled_metadata_field( $proposed, $source_map, 'path', $path, 'git' );
-		$this->set_reconciled_metadata_field( $proposed, $source_map, 'observed_at', gmdate( 'c' ), 'reconcile_run' );
-
-		$origin_site_name = function_exists( 'get_bloginfo' ) ? (string) get_bloginfo( 'name' ) : '';
-		$origin_site_url  = function_exists( 'home_url' ) ? (string) home_url() : '';
-		if ( empty( $proposed['origin_site'] ) ) {
-			$this->set_reconciled_metadata_field( $proposed, $source_map, 'origin_site', '' !== $origin_site_name ? $origin_site_name : $origin_site_url, 'current_site' );
-		}
-		if ( empty( $proposed['origin_site_name'] ) ) {
-			$this->set_reconciled_metadata_field( $proposed, $source_map, 'origin_site_name', $origin_site_name, 'current_site' );
-		}
-		if ( empty( $proposed['origin_site_url'] ) ) {
-			$this->set_reconciled_metadata_field( $proposed, $source_map, 'origin_site_url', $origin_site_url, 'current_site' );
-		}
-
-		$created_source = '';
-		$created_at     = '';
-		foreach ( array( 'created_at', 'timestamp' ) as $key ) {
-			if ( ! empty( $metadata[ $key ] ) && false !== strtotime( (string) $metadata[ $key ] ) ) {
-				$created_at     = gmdate( 'c', (int) strtotime( (string) $metadata[ $key ] ) );
-				$created_source = 'metadata';
-				break;
-			}
-		}
-		if ( '' === $created_at ) {
-			$mtime = file_exists( $path ) ? filemtime( $path ) : false;
-			if ( false !== $mtime ) {
-				$created_at     = gmdate( 'c', (int) $mtime );
-				$created_source = 'filesystem';
-			}
-		}
-		if ( '' !== $created_at ) {
-			$this->set_reconciled_metadata_field( $proposed, $source_map, 'created_at', $created_at, $created_source );
-		}
-
-		$invalid_fields = array();
-		$existing_state = isset( $metadata['lifecycle_state'] ) ? WorktreeContextInjector::normalize_state( (string) $metadata['lifecycle_state'] ) : null;
-		$raw_state      = isset( $metadata['lifecycle_state'] ) ? (string) $metadata['lifecycle_state'] : '';
-		if ( '' !== $raw_state && null === $existing_state ) {
-			$invalid_fields[] = 'lifecycle_state';
-		}
-		if ( null !== $existing_state ) {
-			$this->set_reconciled_metadata_field( $proposed, $source_map, 'lifecycle_state', $existing_state, 'metadata' );
-		} else {
-			$this->set_reconciled_metadata_field( $proposed, $source_map, 'lifecycle_state', WorktreeContextInjector::STATE_ACTIVE, 'operator_plan' );
-		}
-		$missing_after = array_values( array_filter( array(
-			empty( $proposed['created_at'] ) ? 'created_at' : '',
-			empty( $proposed['lifecycle_state'] ) ? 'lifecycle_state' : '',
-		) ) );
-		if ( array() !== $missing_after ) {
-			return array(
-				'skip' => array_merge(
-					$base_row,
-					array(
-						'reason_code'    => 'insufficient_signal',
-						'reason'         => 'not enough stable metadata could be inferred safely',
-						'missing_fields' => $missing_after,
-					)
-				),
-			);
-		}
-
-		$metadata_missing = array();
-		foreach ( array( 'handle', 'repo', 'branch', 'path', 'created_at', 'observed_at', 'lifecycle_state' ) as $field ) {
-			if ( ! array_key_exists( $field, $metadata ) || '' === (string) $metadata[ $field ] ) {
-				$metadata_missing[] = $field;
-			}
-		}
-
-		if ( array() === $metadata_missing && array() === $invalid_fields ) {
-			return array();
-		}
-
-		return array(
-			'proposal' => array_merge(
-				$base_row,
-				array(
-					'reason_code'       => 'metadata_backfill',
-					'reason'            => 'unmanaged worktree metadata can be reconciled without changing cleanup eligibility',
-					'dirty'             => $dirty,
-					'unpushed'          => $unpushed,
-					'missing_fields'    => $metadata_missing,
-					'invalid_fields'    => $invalid_fields,
-					'proposed_metadata' => $proposed,
-					'source_map'        => $source_map,
-				)
-			),
-		);
-	}
-
-	/**
-	 * Check whether metadata already stores the current explicit cleanup state.
-	 *
-	 * Legacy finalized records are still cleanup signals, but reconciliation should
-	 * promote them to explicit cleanup_eligible metadata so later inventory-only
-	 * cleanup has durable evidence to inspect.
-	 *
-	 * @param array<string,mixed> $metadata Worktree metadata.
-	 * @return bool
-	 */
-	private function has_explicit_cleanup_eligible_state( array $metadata ): bool {
-		$state = isset( $metadata['lifecycle_state'] ) ? WorktreeContextInjector::normalize_state( (string) $metadata['lifecycle_state'] ) : null;
-		return WorktreeContextInjector::STATE_CLEANUP_ELIGIBLE === $state;
-	}
-
-	/**
-	 * Detect an unambiguous merge signal for lifecycle reconciliation.
-	 *
-	 * @param array<string,mixed> $wt           Current worktree listing row.
-	 * @param array<string,mixed> $metadata     Persisted lifecycle metadata.
-	 * @param array               $github_cache Run-local GitHub lookup cache.
-	 * @param array               $fetched      Run-local fetched repo cache.
-	 * @return array{signal:string,reason:string,finalized_state?:string,pr_url?:string}|null
-	 */
-	private function detect_worktree_lifecycle_finalizer_signal( array $wt, array $metadata, array &$github_cache, array &$fetched ): ?array {
-		$repo         = (string) ( $wt['repo'] ?? '' );
-		$branch       = (string) ( $wt['branch'] ?? '' );
-		$primary_path = '' !== $repo ? $this->get_primary_path( $repo ) : '';
-		if ( '' === $repo || '' === $branch || ! is_dir( $primary_path . '/.git' ) ) {
-			return null;
-		}
-
-		if ( empty( $fetched[ $repo ] ) ) {
-			$fetch = $this->run_git( $primary_path, 'fetch --prune --quiet origin', self::CLEANUP_GIT_PROBE_TIMEOUT );
-			if ( $this->is_git_timeout_error( $fetch ) ) {
-				return array(
-					'signal' => 'probe-timeout',
-					'reason' => $fetch->get_error_message(),
-				);
-			}
-			$fetched[ $repo ] = true;
-		}
-
-		$pr_signal = $this->detect_stored_pr_merged_signal( $metadata, $github_cache );
-		if ( null !== $pr_signal ) {
-			return $pr_signal;
-		}
-
-		$signal = $this->detect_merge_signal( $primary_path, $repo, $branch, false, $github_cache );
-		if ( null === $signal ) {
-			return null;
-		}
-
-		if ( in_array( (string) ( $signal['signal'] ?? '' ), array( 'upstream-gone', 'local-merged' ), true ) ) {
-			return $signal;
-		}
-
-		return null;
-	}
-
-	/**
-	 * Check stored PR metadata for a merged PR signal.
-	 *
-	 * @param array<string,mixed> $metadata     Persisted lifecycle metadata.
-	 * @param array               $github_cache Run-local GitHub lookup cache.
-	 * @return array{signal:string,reason:string,finalized_state?:string,pr_url?:string}|null
-	 */
-	private function detect_stored_pr_merged_signal( array $metadata, array &$github_cache ): ?array {
-		$pr_repo   = isset( $metadata['pr_repo'] ) ? (string) $metadata['pr_repo'] : '';
-		$pr_number = isset( $metadata['pr_number'] ) ? (int) $metadata['pr_number'] : 0;
-		$pr_url    = isset( $metadata['pr_url'] ) ? (string) $metadata['pr_url'] : '';
-
-		if ( ( '' === $pr_repo || $pr_number <= 0 ) && '' !== $pr_url && preg_match( '~^https?://github\.com/([^/]+)/([^/]+)/pull/(\d+)(?:[/?#].*)?$~', $pr_url, $matches ) ) {
-			$pr_repo   = $matches[1] . '/' . $matches[2];
-			$pr_number = (int) $matches[3];
-		}
-
-		if ( '' === $pr_repo || $pr_number <= 0 ) {
-			return null;
-		}
-
-		$cache_key = $pr_repo . '#' . $pr_number;
-		if ( array_key_exists( $cache_key, $github_cache ) ) {
-			$pr = $github_cache[ $cache_key ];
-		} else {
-			$pr = $this->fetch_github_pull_request( $pr_repo, $pr_number );
-			$github_cache[ $cache_key ] = $pr;
-		}
-
-		if ( ! is_array( $pr ) ) {
-			return null;
-		}
-
-		$state = (string) ( $pr['state'] ?? '' );
-		if ( empty( $pr['merged_at'] ) && 'closed' !== $state ) {
-			return null;
-		}
-
-		$merged = ! empty( $pr['merged_at'] );
-
-		return array(
-			'signal'          => $merged ? 'pr-merged' : 'pr-closed',
-			'reason'          => $merged ? sprintf( 'stored PR #%d merged (%s)', $pr_number, $state ) : sprintf( 'stored PR #%d closed without merge', $pr_number ),
-			'finalized_state' => $merged ? WorktreeContextInjector::STATE_MERGED : WorktreeContextInjector::STATE_CLOSED,
-			'pr_url'          => (string) ( $pr['html_url'] ?? $pr_url ),
-		);
-	}
-
-	/**
-	 * Fetch one pull request from GitHub when API credentials are available.
-	 *
-	 * @return array<string,mixed>|null
-	 */
-	private function fetch_github_pull_request( string $slug, int $number ): ?array {
-		if ( ! class_exists( '\DataMachineCode\Abilities\GitHubAbilities' ) ) {
-			return null;
-		}
-
-		// Pass the repo through so credential profiles with `allowed_repos`
-		// can win over the global default profile when fetching merge state.
-		$pat = \DataMachineCode\Abilities\GitHubAbilities::getPat( array( 'repo' => $slug ) );
-		if ( empty( $pat ) ) {
-			return null;
-		}
-
-		$response = \DataMachineCode\Abilities\GitHubAbilities::apiGet(
-			GitHubRemote::apiUrl( $slug, 'pulls/' . $number ),
-			array(),
-			$pat,
-			self::CLEANUP_GITHUB_TIMEOUT
-		);
-		if ( is_wp_error( $response ) ) {
-			return null;
-		}
-
-		$data = $response['data'] ?? null;
-		return is_array( $data ) ? $data : null;
-	}
-
-	/**
-	 * Set a reconciled metadata field and record its source.
-	 *
-	 * @param array<string,mixed> $metadata Metadata under construction.
-	 * @param array<string,string> $source_map Field source map.
-	 * @param string $field Field name.
-	 * @param mixed  $value Field value.
-	 * @param string $source Source label.
-	 */
-	private function set_reconciled_metadata_field( array &$metadata, array &$source_map, string $field, mixed $value, string $source ): void {
-		if ( null === $value || '' === (string) $value ) {
-			return;
-		}
-		$metadata[ $field ]    = $value;
-		$source_map[ $field ]  = $source;
-	}
-
-	/**
-	 * Apply a reviewed metadata reconciliation plan after exact revalidation.
-	 *
-	 * @param array<string,mixed> $plan Dry-run plan.
-	 * @return array<string,mixed>|\WP_Error
-	 */
-	private function apply_worktree_metadata_reconciliation_plan( array $plan ): array|\WP_Error {
-		$planned = $this->extract_worktree_metadata_reconciliation_plan( $plan );
-		if ( $planned instanceof \WP_Error ) {
-			return $planned;
-		}
-
-		$current_by_handle = array();
-		if ( ! empty( $plan['direct_apply'] ) ) {
-			foreach ( $planned as $row ) {
-				$handle = (string) ( $row['handle'] ?? '' );
-				if ( '' === $handle ) {
-					continue;
-				}
-				$current_by_handle[ $handle ] = array(
-					'handle' => $handle,
-					'repo'   => (string) ( $row['repo'] ?? '' ),
-					'branch' => (string) ( $row['branch'] ?? '' ),
-					'path'   => (string) ( $row['path'] ?? '' ),
-					'dirty'  => (int) ( $row['dirty'] ?? 0 ),
-				);
-			}
-		} else {
-			$listing = $this->worktree_list();
-			if ( $listing instanceof \WP_Error ) {
-				return $listing;
-			}
-
-			foreach ( (array) ( $listing['worktrees'] ?? array() ) as $wt ) {
-				$handle = (string) ( $wt['handle'] ?? '' );
-				if ( '' !== $handle ) {
-					$current_by_handle[ $handle ] = $wt;
-				}
-			}
-		}
-
-		$written = array();
-		$skipped = array();
-		foreach ( $planned as $row ) {
-			$handle  = (string) ( $row['handle'] ?? '' );
-			$current = $current_by_handle[ $handle ] ?? null;
-			if ( null === $current ) {
-				$skipped[] = $this->build_reconcile_apply_skip( $row, null, 'plan_not_current', 'planned worktree is no longer present in the current listing' );
-				continue;
-			}
-
-			$mismatches = array();
-			foreach ( array( 'repo', 'branch', 'path' ) as $field ) {
-				if ( (string) ( $row[ $field ] ?? '' ) !== (string) ( $current[ $field ] ?? '' ) ) {
-					$mismatches[] = sprintf( '%s planned=%s current=%s', $field, (string) ( $row[ $field ] ?? '' ), (string) ( $current[ $field ] ?? '' ) );
-				}
-			}
-			if ( array() !== $mismatches ) {
-				$skipped[] = $this->build_reconcile_apply_skip( $row, $current, 'plan_identity_mismatch', 'planned identity does not match current row: ' . implode( '; ', $mismatches ) );
-				continue;
-			}
-
-			$metadata   = (array) ( $row['proposed_metadata'] ?? array() );
-			$source_map = (array) ( $row['source_map'] ?? array() );
-			$state      = WorktreeContextInjector::normalize_state( (string) ( $metadata['lifecycle_state'] ?? '' ) );
-			if ( null === $state ) {
-				$skipped[] = $this->build_reconcile_apply_skip( $row, $current, 'invalid_lifecycle_state', 'proposed lifecycle_state is invalid' );
-				continue;
-			}
-			$metadata['lifecycle_state'] = $state;
-
-			if ( WorktreeContextInjector::STATE_CLEANUP_ELIGIBLE === $state ) {
-				$path     = (string) ( $current['path'] ?? '' );
-				$dirty    = ! empty( $plan['direct_apply'] ) ? $this->probe_worktree_dirty_count( $path, self::CLEANUP_GIT_PROBE_TIMEOUT ) : (int) ( $current['dirty'] ?? 0 );
-				if ( is_wp_error( $dirty ) ) {
-					$skipped[] = $this->build_reconcile_apply_skip( $row, $current, 'probe_timeout', 'dirty-state probe failed - refusing cleanup_eligible metadata write: ' . $dirty->get_error_message() );
-					continue;
-				}
-				$unpushed = $this->count_unpushed_commits( $path );
-				if ( $dirty > 0 || $unpushed > 0 ) {
-					$skipped[] = $this->build_reconcile_apply_skip( $row, $current, 'unsafe_cleanup_eligible_state', 'refusing to mark dirty or unpushed worktree cleanup_eligible from a reconciliation plan' );
-					continue;
-				}
-			}
-
-			foreach ( array( 'handle', 'repo', 'branch', 'path', 'created_at', 'observed_at', 'lifecycle_state' ) as $field ) {
-				if ( ! isset( $source_map[ $field ] ) || '' === (string) $source_map[ $field ] ) {
-					$skipped[] = $this->build_reconcile_apply_skip( $row, $current, 'missing_source_map', sprintf( 'proposed field %s is missing a source', $field ) );
-					continue 2;
-				}
-			}
-
-			$metadata['reconciled_at']      = gmdate( 'c' );
-			$metadata['reconciled_sources'] = $source_map;
-			WorktreeContextInjector::store_lifecycle_metadata( $handle, $metadata );
-
-			$written[] = array(
-				'handle'   => $handle,
-				'repo'     => (string) ( $current['repo'] ?? '' ),
-				'branch'   => (string) ( $current['branch'] ?? '' ),
-				'path'     => (string) ( $current['path'] ?? '' ),
-				'metadata' => WorktreeContextInjector::get_metadata( $handle ),
-			);
-		}
-
-		$classified_skips = $this->classify_worktree_metadata_reconciliation_skips( $skipped );
-
-		$inspected = isset( $plan['summary']['inspected'] ) ? (int) $plan['summary']['inspected'] : count( (array) ( $listing['worktrees'] ?? array() ) );
-		$result    = array(
-			'success'        => true,
-			'dry_run'        => false,
-			'applied'        => true,
-			'direct_apply'   => ! empty( $plan['direct_apply'] ),
-			'generated_at'   => gmdate( 'c' ),
-			'workspace_path' => $this->workspace_path,
-			'proposals'      => $planned,
-			'written'        => $written,
-			'skipped'        => $skipped,
-			'still_unsafe'      => $classified_skips['still_unsafe'],
-			'external_worktrees' => $classified_skips['external_worktrees'],
-			'summary'        => $this->build_worktree_metadata_reconciliation_summary( $inspected, $planned, $written, $skipped ),
-		);
-
-		if ( isset( $plan['pagination'] ) && is_array( $plan['pagination'] ) ) {
-			$result['pagination'] = $plan['pagination'];
-		}
-		if ( isset( $plan['evidence'] ) && is_array( $plan['evidence'] ) ) {
-			$result['evidence'] = array_merge(
-				$plan['evidence'],
-				array(
-					'scope'        => ! empty( $plan['direct_apply'] ) ? 'paginated metadata reconciliation direct apply' : (string) ( $plan['evidence']['scope'] ?? 'paginated metadata reconciliation apply-plan' ),
-					'apply_source' => ! empty( $plan['direct_apply'] ) ? 'direct_apply' : 'apply_plan',
-				)
-			);
-		}
-
-		return $result;
-	}
-
-	/**
-	 * Split reconciliation skips into stable operator-facing buckets.
-	 *
-	 * @param array<int,array<string,mixed>> $skipped Skipped rows.
-	 * @return array{still_unsafe:array<int,array<string,mixed>>,external_worktrees:array<int,array<string,mixed>>}
-	 */
-	private function classify_worktree_metadata_reconciliation_skips( array $skipped ): array {
-		$still_unsafe       = array();
-		$external_worktrees = array();
-
-		foreach ( $skipped as $row ) {
-			$reason_code = (string) ( $row['reason_code'] ?? '' );
-			if ( 'external_worktree' === $reason_code ) {
-				$external_worktrees[] = $row;
-			}
-			if ( in_array( $reason_code, array( 'unsafe_cleanup_eligible_state', 'plan_identity_mismatch', 'plan_not_current' ), true ) ) {
-				$still_unsafe[] = $row;
-			}
-		}
-
-		return array(
-			'still_unsafe'      => $still_unsafe,
-			'external_worktrees' => $external_worktrees,
-		);
-	}
-
-	/**
-	 * Extract reconciliation proposals from a dry-run plan.
-	 *
-	 * @param array<string,mixed> $plan Plan file.
-	 * @return array<int,array<string,mixed>>|\WP_Error
-	 */
-	private function extract_worktree_metadata_reconciliation_plan( array $plan ): array|\WP_Error {
-		$proposals = $plan['proposals'] ?? null;
-		if ( ! is_array( $proposals ) ) {
-			return new \WP_Error( 'invalid_metadata_reconcile_plan', 'Metadata reconciliation plan must contain a proposals array.', array( 'status' => 400 ) );
-		}
-
-		foreach ( $proposals as $index => $row ) {
-			if ( ! is_array( $row ) ) {
-				return new \WP_Error( 'invalid_metadata_reconcile_plan', sprintf( 'Plan proposal #%d is not an object.', (int) $index ), array( 'status' => 400 ) );
-			}
-			foreach ( array( 'handle', 'repo', 'branch', 'path', 'proposed_metadata', 'source_map' ) as $field ) {
-				if ( ! array_key_exists( $field, $row ) || ( is_string( $row[ $field ] ) && '' === trim( $row[ $field ] ) ) ) {
-					return new \WP_Error( 'invalid_metadata_reconcile_plan', sprintf( 'Plan proposal #%d is missing %s.', (int) $index, $field ), array( 'status' => 400 ) );
-				}
-			}
-		}
-
-		return array_values( $proposals );
-	}
-
-	/**
-	 * Build an apply skip row.
-	 *
-	 * @param array<string,mixed>      $planned Planned row.
-	 * @param array<string,mixed>|null $current Current row.
-	 * @param string                   $code Reason code.
-	 * @param string                   $reason Human reason.
-	 * @return array<string,mixed>
-	 */
-	private function build_reconcile_apply_skip( array $planned, ?array $current, string $code, string $reason ): array {
-		return array(
-			'handle'      => (string) ( $planned['handle'] ?? '' ),
-			'repo'        => (string) ( $current['repo'] ?? $planned['repo'] ?? '' ),
-			'branch'      => (string) ( $current['branch'] ?? $planned['branch'] ?? '' ),
-			'path'        => (string) ( $current['path'] ?? $planned['path'] ?? '' ),
-			'reason_code' => $code,
-			'reason'      => $reason,
-			'planned'     => $planned,
-			'current'     => $current,
-		);
-	}
-
-	/**
-	 * Build metadata reconciliation pagination evidence.
-	 *
-	 * @return array<string,mixed>
-	 */
-	private function build_worktree_metadata_reconciliation_pagination( int $total, int $scanned, int $limit, int $offset ): array {
-		$next_offset = $offset + $scanned;
-		$complete    = $next_offset >= $total;
-
-		return array(
-			'total'       => $total,
-			'offset'      => $offset,
-			'limit'       => $limit,
-			'scanned'     => $scanned,
-			'partial'     => ! $complete,
-			'complete'    => $complete,
-			'next_offset' => $complete ? null : $next_offset,
-		);
-	}
-
-	/**
-	 * Probe the dirty file count for a single worktree path.
-	 *
-	 * @param string $path Worktree path.
-	 * @return int|\WP_Error Dirty file count, or WP_Error when git probe failed.
-	 */
-	private function probe_worktree_dirty_count( string $path, int $timeout_seconds = 0 ): int|\WP_Error {
-		if ( '' === $path || ! is_dir( $path ) ) {
-			return new \WP_Error( 'worktree_path_missing', 'worktree path is not a directory', array( 'status' => 400 ) );
-		}
-		$result = $this->run_git( $path, 'status --porcelain', $timeout_seconds );
-		if ( is_wp_error( $result ) ) {
-			return $result;
-		}
-		$lines = array_filter( array_map( 'trim', explode( "\n", (string) ( $result['output'] ?? '' ) ) ) );
-		return count( $lines );
-	}
-
-	/**
-	 * Build stable reconciliation summary counts.
-	 *
-	 * @param int   $inspected Worktree rows inspected.
-	 * @param array $proposals Proposal rows.
-	 * @param array $written Written rows.
-	 * @param array $skipped Skipped rows.
-	 * @return array<string,mixed>
-	 */
-	private function build_worktree_metadata_reconciliation_summary( int $inspected, array $proposals, array $written, array $skipped ): array {
-		$skipped_by_reason = array();
-		foreach ( $skipped as $row ) {
-			$code                       = (string) ( $row['reason_code'] ?? 'unknown' );
-			$skipped_by_reason[ $code ] = ( $skipped_by_reason[ $code ] ?? 0 ) + 1;
-		}
-		ksort( $skipped_by_reason );
-
-		$states = array();
-		foreach ( $proposals as $row ) {
-			$state            = (string) ( $row['proposed_metadata']['lifecycle_state'] ?? 'unknown' );
-			$states[ $state ] = ( $states[ $state ] ?? 0 ) + 1;
-		}
-		ksort( $states );
-
-		return array(
-			'inspected'         => $inspected,
-			'proposed'          => count( $proposals ),
-			'written'           => count( $written ),
-			'skipped'           => count( $skipped ),
-			'skipped_by_reason' => $skipped_by_reason,
-			'proposed_by_state' => $states,
-			'slow_rows'         => $this->summarize_slow_worktree_rows( array_merge( $proposals, $skipped ) ),
-		);
-	}
-
-	/**
-	 * Summarize the slowest worktree rows from an expensive report page.
-	 *
-	 * @param array<int,array<string,mixed>> $rows Timed rows.
-	 * @return array<int,array<string,mixed>>
-	 */
-	private function summarize_slow_worktree_rows( array $rows ): array {
-		$timed = array_values( array_filter( $rows, fn( $row ) => is_array( $row ) && isset( $row['elapsed_ms'] ) ) );
-		usort( $timed, fn( $a, $b ) => (int) ( $b['elapsed_ms'] ?? 0 ) <=> (int) ( $a['elapsed_ms'] ?? 0 ) );
-
-		return array_map(
-			fn( $row ) => array_filter(
-				array(
-					'handle'      => (string) ( $row['handle'] ?? '' ),
-					'repo'        => (string) ( $row['repo'] ?? '' ),
-					'branch'      => (string) ( $row['branch'] ?? '' ),
-					'elapsed_ms'  => (int) ( $row['elapsed_ms'] ?? 0 ),
-					'action'      => (string) ( $row['suggested_action'] ?? '' ),
-					'reason_code' => (string) ( $row['reason_code'] ?? '' ),
-				),
-				fn( $value ) => '' !== $value
-			),
-			array_slice( $timed, 0, 10 )
 		);
 	}
